chore: move wo critical table to app
