+++ src/apps/Construction/Components/DetailsPage/components/footers/SumColumnFooterCountTotal.tsx
