+++ src/apps/Construction/Components/DetailsPage/components/footers/index.ts
