+++ src/apps/Construction/Components/DetailsPage/components/index.ts
