+++ src/apps/Construction/Components/DetailsPage/tableConfig.tsx
@@ -1,5 +1,5 @@
 import { Column } from '@equinor/Table';
-import { WorkOrder } from '../mocData/mockData';
+import { WorkOrder } from '../../mocData/mockData';
 import { SumColumnFooter } from './components';
 import { columnGenerator } from './utils';
