+++ src/apps/Construction/Components/DetailsPage/types.ts
