+++ src/apps/Construction/Components/index.ts
@@ -1,0 +1,2 @@
+export * from './CriticalWoTable';
+export * from './DetailsPage';
